Added states
You can now change the appearence(and maybe more in the future) of your boxes with the state of the box. The css classes of the states are listed in config.states.

The update function of the custom box now returns an array with the html and the state in it. You have to change this in your update functions, as it will break otherwise.

I've also fixed the .content property of the box not being updated.